LayoutTests/fast/selectors/nth-child-of-chained-3-expected.txt

Test the chaining of multiple :nth-child(An+B of selector) does match every one of the selector.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing "testcaseA.foo[id]"
PASS document.querySelectorAll('#test-root testcaseA.foo[id]').length is 12
PASS document.querySelectorAll('#test-root testcaseA.foo[id]')[0].id is "testcase3"
PASS document.querySelectorAll('#test-root testcaseA.foo[id]')[1].id is "testcase5"
PASS document.querySelectorAll('#test-root testcaseA.foo[id]')[2].id is "testcase9"
PASS document.querySelectorAll('#test-root testcaseA.foo[id]')[3].id is "testcase11"
PASS document.querySelectorAll('#test-root testcaseA.foo[id]')[4].id is "testcase13"
PASS document.querySelectorAll('#test-root testcaseA.foo[id]')[5].id is "testcase15"
PASS document.querySelectorAll('#test-root testcaseA.foo[id]')[6].id is "testcase19"
PASS document.querySelectorAll('#test-root testcaseA.foo[id]')[7].id is "testcase21"
PASS document.querySelectorAll('#test-root testcaseA.foo[id]')[8].id is "testcase23"
PASS document.querySelectorAll('#test-root testcaseA.foo[id]')[9].id is "testcase25"
PASS document.querySelectorAll('#test-root testcaseA.foo[id]')[10].id is "testcase27"
PASS document.querySelectorAll('#test-root testcaseA.foo[id]')[11].id is "testcase29"
PASS styledElements is ["testcase3","testcase5","testcase9","testcase11","testcase13","testcase15","testcase19","testcase21","testcase23","testcase25","testcase27","testcase29"]

Testing ":nth-child(n of .foo):nth-child(n of testcaseA):nth-child(n of [id])"
PASS document.querySelectorAll('#test-root :nth-child(n of .foo):nth-child(n of testcaseA):nth-child(n of [id])').length is 12
PASS document.querySelectorAll('#test-root :nth-child(n of .foo):nth-child(n of testcaseA):nth-child(n of [id])')[0].id is "testcase3"
PASS document.querySelectorAll('#test-root :nth-child(n of .foo):nth-child(n of testcaseA):nth-child(n of [id])')[1].id is "testcase5"
PASS document.querySelectorAll('#test-root :nth-child(n of .foo):nth-child(n of testcaseA):nth-child(n of [id])')[2].id is "testcase9"
PASS document.querySelectorAll('#test-root :nth-child(n of .foo):nth-child(n of testcaseA):nth-child(n of [id])')[3].id is "testcase11"
PASS document.querySelectorAll('#test-root :nth-child(n of .foo):nth-child(n of testcaseA):nth-child(n of [id])')[4].id is "testcase13"
PASS document.querySelectorAll('#test-root :nth-child(n of .foo):nth-child(n of testcaseA):nth-child(n of [id])')[5].id is "testcase15"
PASS document.querySelectorAll('#test-root :nth-child(n of .foo):nth-child(n of testcaseA):nth-child(n of [id])')[6].id is "testcase19"
PASS document.querySelectorAll('#test-root :nth-child(n of .foo):nth-child(n of testcaseA):nth-child(n of [id])')[7].id is "testcase21"
PASS document.querySelectorAll('#test-root :nth-child(n of .foo):nth-child(n of testcaseA):nth-child(n of [id])')[8].id is "testcase23"
PASS document.querySelectorAll('#test-root :nth-child(n of .foo):nth-child(n of testcaseA):nth-child(n of [id])')[9].id is "testcase25"
PASS document.querySelectorAll('#test-root :nth-child(n of .foo):nth-child(n of testcaseA):nth-child(n of [id])')[10].id is "testcase27"
PASS document.querySelectorAll('#test-root :nth-child(n of .foo):nth-child(n of testcaseA):nth-child(n of [id])')[11].id is "testcase29"
PASS styledElements is ["testcase3","testcase5","testcase9","testcase11","testcase13","testcase15","testcase19","testcase21","testcase23","testcase25","testcase27","testcase29"]

Testing ":nth-child(n+1 of .foo):nth-child(n of testcaseA):nth-child(n+1 of [id])"
PASS document.querySelectorAll('#test-root :nth-child(n+1 of .foo):nth-child(n of testcaseA):nth-child(n+1 of [id])').length is 12
PASS document.querySelectorAll('#test-root :nth-child(n+1 of .foo):nth-child(n of testcaseA):nth-child(n+1 of [id])')[0].id is "testcase3"
PASS document.querySelectorAll('#test-root :nth-child(n+1 of .foo):nth-child(n of testcaseA):nth-child(n+1 of [id])')[1].id is "testcase5"
PASS document.querySelectorAll('#test-root :nth-child(n+1 of .foo):nth-child(n of testcaseA):nth-child(n+1 of [id])')[2].id is "testcase9"
PASS document.querySelectorAll('#test-root :nth-child(n+1 of .foo):nth-child(n of testcaseA):nth-child(n+1 of [id])')[3].id is "testcase11"
PASS document.querySelectorAll('#test-root :nth-child(n+1 of .foo):nth-child(n of testcaseA):nth-child(n+1 of [id])')[4].id is "testcase13"
PASS document.querySelectorAll('#test-root :nth-child(n+1 of .foo):nth-child(n of testcaseA):nth-child(n+1 of [id])')[5].id is "testcase15"
PASS document.querySelectorAll('#test-root :nth-child(n+1 of .foo):nth-child(n of testcaseA):nth-child(n+1 of [id])')[6].id is "testcase19"
PASS document.querySelectorAll('#test-root :nth-child(n+1 of .foo):nth-child(n of testcaseA):nth-child(n+1 of [id])')[7].id is "testcase21"
PASS document.querySelectorAll('#test-root :nth-child(n+1 of .foo):nth-child(n of testcaseA):nth-child(n+1 of [id])')[8].id is "testcase23"
PASS document.querySelectorAll('#test-root :nth-child(n+1 of .foo):nth-child(n of testcaseA):nth-child(n+1 of [id])')[9].id is "testcase25"
PASS document.querySelectorAll('#test-root :nth-child(n+1 of .foo):nth-child(n of testcaseA):nth-child(n+1 of [id])')[10].id is "testcase27"
PASS document.querySelectorAll('#test-root :nth-child(n+1 of .foo):nth-child(n of testcaseA):nth-child(n+1 of [id])')[11].id is "testcase29"
PASS styledElements is ["testcase3","testcase5","testcase9","testcase11","testcase13","testcase15","testcase19","testcase21","testcase23","testcase25","testcase27","testcase29"]

Testing ":nth-child(n+1 of .foo):nth-child(n+1 of testcaseA):nth-child(n of [id])"
PASS document.querySelectorAll('#test-root :nth-child(n+1 of .foo):nth-child(n+1 of testcaseA):nth-child(n of [id])').length is 12
PASS document.querySelectorAll('#test-root :nth-child(n+1 of .foo):nth-child(n+1 of testcaseA):nth-child(n of [id])')[0].id is "testcase3"
PASS document.querySelectorAll('#test-root :nth-child(n+1 of .foo):nth-child(n+1 of testcaseA):nth-child(n of [id])')[1].id is "testcase5"
PASS document.querySelectorAll('#test-root :nth-child(n+1 of .foo):nth-child(n+1 of testcaseA):nth-child(n of [id])')[2].id is "testcase9"
PASS document.querySelectorAll('#test-root :nth-child(n+1 of .foo):nth-child(n+1 of testcaseA):nth-child(n of [id])')[3].id is "testcase11"
PASS document.querySelectorAll('#test-root :nth-child(n+1 of .foo):nth-child(n+1 of testcaseA):nth-child(n of [id])')[4].id is "testcase13"
PASS document.querySelectorAll('#test-root :nth-child(n+1 of .foo):nth-child(n+1 of testcaseA):nth-child(n of [id])')[5].id is "testcase15"
PASS document.querySelectorAll('#test-root :nth-child(n+1 of .foo):nth-child(n+1 of testcaseA):nth-child(n of [id])')[6].id is "testcase19"
PASS document.querySelectorAll('#test-root :nth-child(n+1 of .foo):nth-child(n+1 of testcaseA):nth-child(n of [id])')[7].id is "testcase21"
PASS document.querySelectorAll('#test-root :nth-child(n+1 of .foo):nth-child(n+1 of testcaseA):nth-child(n of [id])')[8].id is "testcase23"
PASS document.querySelectorAll('#test-root :nth-child(n+1 of .foo):nth-child(n+1 of testcaseA):nth-child(n of [id])')[9].id is "testcase25"
PASS document.querySelectorAll('#test-root :nth-child(n+1 of .foo):nth-child(n+1 of testcaseA):nth-child(n of [id])')[10].id is "testcase27"
PASS document.querySelectorAll('#test-root :nth-child(n+1 of .foo):nth-child(n+1 of testcaseA):nth-child(n of [id])')[11].id is "testcase29"
PASS styledElements is ["testcase3","testcase5","testcase9","testcase11","testcase13","testcase15","testcase19","testcase21","testcase23","testcase25","testcase27","testcase29"]

Testing ":nth-child(n+1 of .foo):nth-child(2n of testcaseA):nth-child(n of [id])"
PASS document.querySelectorAll('#test-root :nth-child(n+1 of .foo):nth-child(2n of testcaseA):nth-child(n of [id])').length is 6
PASS document.querySelectorAll('#test-root :nth-child(n+1 of .foo):nth-child(2n of testcaseA):nth-child(n of [id])')[0].id is "testcase3"
PASS document.querySelectorAll('#test-root :nth-child(n+1 of .foo):nth-child(2n of testcaseA):nth-child(n of [id])')[1].id is "testcase11"
PASS document.querySelectorAll('#test-root :nth-child(n+1 of .foo):nth-child(2n of testcaseA):nth-child(n of [id])')[2].id is "testcase15"
PASS document.querySelectorAll('#test-root :nth-child(n+1 of .foo):nth-child(2n of testcaseA):nth-child(n of [id])')[3].id is "testcase19"
PASS document.querySelectorAll('#test-root :nth-child(n+1 of .foo):nth-child(2n of testcaseA):nth-child(n of [id])')[4].id is "testcase23"
PASS document.querySelectorAll('#test-root :nth-child(n+1 of .foo):nth-child(2n of testcaseA):nth-child(n of [id])')[5].id is "testcase27"
PASS styledElements is ["testcase3","testcase11","testcase15","testcase19","testcase23","testcase27"]

Testing ":nth-child(2n+1 of .foo):nth-child(2n of testcaseA):nth-child(n of [id])"
PASS document.querySelectorAll('#test-root :nth-child(2n+1 of .foo):nth-child(2n of testcaseA):nth-child(n of [id])').length is 3
PASS document.querySelectorAll('#test-root :nth-child(2n+1 of .foo):nth-child(2n of testcaseA):nth-child(n of [id])')[0].id is "testcase3"
PASS document.querySelectorAll('#test-root :nth-child(2n+1 of .foo):nth-child(2n of testcaseA):nth-child(n of [id])')[1].id is "testcase15"
PASS document.querySelectorAll('#test-root :nth-child(2n+1 of .foo):nth-child(2n of testcaseA):nth-child(n of [id])')[2].id is "testcase27"
PASS styledElements is ["testcase3","testcase15","testcase27"]

PASS successfullyParsed is true

TEST COMPLETE
